+++ data-raw/Sentinel_2_Spectral_Response.R
@@ -1,7 +1,7 @@
 ## code to prepare `Sentinel_2` spectral response
 # 1- define sensor name
 SensorName <-'Sentinel_2'
-Path_SRF <- file.path('data-raw',paste(SensorName,'_Spectral_Response.csv',sep=''))
+Path_SRF <- file.path('data-raw',paste0(SensorName,'_Spectral_Response.csv'))
 # 2- read file containing spectral response
 message('_____ reading spectral response corresponding to ______')
 print(SensorName)
@@ -15,7 +15,8 @@ Spectral_Bands <- Spectral_Bands[-1]
 # 5- check if conversion of spctral bands into numeric values
 SensorRadiometry <-SRFraw[,-1]
 SensorRadiometry <- t(SensorRadiometry)
-Central_WL <- c(492.4, 559.8, 664.6, 704.1, 740.5, 782.8, 832.8, 864.7, 1613.7, 2202.4)
+Central_WL <- c(492.4, 559.8, 664.6, 704.1, 740.5, 782.8, 832.8, 864.7,
+                1613.7, 2202.4)
 Sentinel_2 <- list('Spectral_Response' = SensorRadiometry,
                    'Spectral_Bands' = Spectral_Bands,
                    'Original_Bands' = Original_Bands,
@@ -23,4 +24,4 @@ Sentinel_2 <- list('Spectral_Response' = SensorRadiometry,
 
 ## code to prepare `Sentinel_2_Spectral_Response` dataset goes here
 usethis::use_data(Sentinel_2,compress = 'xz',overwrite = TRUE)
-save(Sentinel_2,file =  file.path('data',paste(SensorName,'.RData',sep = '')))
+save(Sentinel_2,file =  file.path('data',paste0(SensorName,'.RData')))